+++ comments-bundle/src/Resources/contao/models/CommentsNotifyModel.php
@@ -124,6 +124,20 @@ class CommentsNotifyModel extends Model
 
 		return static::findBy(array("$t.source=? AND $t.parent=? AND tokenConfirm=''"), array($strSource, $intParent), $arrOptions);
 	}
+
+	/**
+	 * Find subscriptions that have not been activated for more than 24 hours
+	 *
+	 * @param array $arrOptions An optional options array
+	 *
+	 * @return Model\Collection|CommentsNotifyModel[]|CommentsNotifyModel|null A collection of models or null if there are no active subscriptions
+	 */
+	public static function findExpiredSubscriptions(array $arrOptions=array())
+	{
+		$t = static::$strTable;
+
+		return static::findBy(array("$t.addedOn<? AND $t.tokenConfirm!=''"), array(strtotime('-1 day')), $arrOptions);
+	}
 }
 
 class_alias(CommentsNotifyModel::class, 'CommentsNotifyModel');