+++ comments-bundle/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## DEV
 
+ * Remove subscriptions that are not activated within 24 hours (see contao/core-bundle#1512).
  * Append the parent ID to the form field IDs to prevent duplicate IDs (see contao/core-bundle#1493).
  * Replace `<div class="quote">` with `<blockquote>` (see contao/core#2244).